Update task status in the executor
The task scheduler relies on the task status to schedule tasks.
Updating the task status should be done as close to the task
execution as possible. Instead of updating the task status in
the callback method, which relies on the executor futures
implementation it is moved to the task executor thread. This
protects against the risk (albeit slight risk) that the futures
thread isn't launched correctly.

Signed-off-by: Tommy Shiou <tshiou@turn.com>