Prevent empty state by reordering state updates
After the execution of a task the task status should be updated
and the in-progress status should be removed. To avoid any
brief window of time where not status is committed for the task,
the task execution result status (SUCCESS or ERROR) should be
committed first before removing the IN_PROGRESS status.

Signed-off-by: Tommy Shiou <tshiou@turn.com>